==> BACKEND/notebooks/01_data_exploration.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

pd.set_option("display.max_columns", None)



data = pd.read_csv("../data/diabetic_data.csv")
ids = pd.read_csv("../data/IDS_mapping.csv")



data.shape

data.head()

data.info()


(data == '?').sum().sort_values(ascending=False).head(20)



data['readmitted'].value_counts(normalize=True)
