==> BACKEND/notebooks/02_preprocessing.ipynb <==
import pandas as pd
import numpy as np




data = pd.read_csv("../data/diabetic_data.csv")



data.shape



df = data.copy()



df.drop(columns=['encounter_id', 'patient_nbr'], inplace=True)



df.replace('?', np.nan, inplace=True)



age_map = {
    '[0-10)': 5, '[10-20)': 15, '[20-30)': 25,
    '[30-40)': 35, '[40-50)': 45, '[50-60)': 55,
    '[60-70)': 65, '[70-80)': 75, '[80-90)': 85,
    '[90-100)': 95
}

df['age'] = df['age'].map(age_map)


df.head()
df['age'].isna().sum()


# Save cleaned data for next stage
df.to_csv("../data/processed_diabetic_data.csv", index=False)

print("✅ Preprocessed data saved successfully")
