==> BACKEND/notebooks/03_feature_engineering.ipynb <==
import pandas as pd
import numpy as np



df = pd.read_csv("../data/processed_diabetic_data.csv")

df.head()



df['target'] = np.where(df['readmitted'] == '<30', 1, 0)

df['target'].value_counts()


features = [
    'age',
    'gender',
    'time_in_hospital',
    'num_lab_procedures',
    'num_medications',
    'number_outpatient',
    'number_emergency',
    'number_inpatient',
    'number_diagnoses',
    'insulin',
    'diabetesMed'
]

X = df[features]
y = df['target']


X = pd.get_dummies(X, drop_first=True)

X.shape


from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.2,
    stratify=y,
    random_state=42
)


X_train.to_csv("../data/X_train.csv", index=False)
X_test.to_csv("../data/X_test.csv", index=False)
y_train.to_csv("../data/y_train.csv", index=False)
y_test.to_csv("../data/y_test.csv", index=False)

print("✅ Feature engineering completed")
