==> BACKEND/notebooks/04_model_training.ipynb <==
import pandas as pd
import numpy as np



# Load feature-engineered data
X_train = pd.read_csv("../data/X_train.csv")
X_test  = pd.read_csv("../data/X_test.csv")
y_train = pd.read_csv("../data/y_train.csv").values.ravel()
y_test  = pd.read_csv("../data/y_test.csv").values.ravel()

print(X_train.shape, y_train.shape)


from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)


from sklearn.neural_network import MLPClassifier

model = MLPClassifier(
    hidden_layer_sizes=(128, 64, 32),
    activation='relu',
    solver='adam',
    max_iter=50,
    random_state=42,
    verbose=True
)

model.fit(X_train_scaled, y_train)


from sklearn.metrics import accuracy_score, classification_report, roc_auc_score

y_pred = model.predict(X_test_scaled)
y_proba = model.predict_proba(X_test_scaled)[:, 1]

print("Accuracy:", accuracy_score(y_test, y_pred))
print("ROC AUC:", roc_auc_score(y_test, y_proba))
print(classification_report(y_test, y_pred))


import joblib

joblib.dump(model, "../model/ann_model.pkl")
joblib.dump(scaler, "../model/scaler.pkl")
joblib.dump(list(X_train.columns), "../model/feature_names.pkl")

print("✅ Model artifacts saved successfully")
